LPS-53617 RSSAction: move logic to JournalRSSUtil
