+++ portal-impl/src/com/liferay/portlet/journal/util/JournalRSSUtil.java
@@ -341,4 +341,306 @@ public class JournalRSSUtil {
 
 	private static final Log _log = LogFactoryUtil.getLog(JournalRSSUtil.class);
 
+	protected String exportToRSS(
+			ResourceRequest resourceRequest, ResourceResponse resourceResponse,
+			JournalFeed feed, String languageId, Layout layout,
+			ThemeDisplay themeDisplay)
+		throws Exception {
+
+		SyndFeed syndFeed = new SyndFeedImpl();
+
+		syndFeed.setDescription(feed.getDescription());
+
+		List<SyndEntry> syndEntries = new ArrayList<>();
+
+		syndFeed.setEntries(syndEntries);
+
+		List<JournalArticle> articles = com.liferay.portlet.journal.util.JournalRSSUtil.getArticles(feed);
+
+		if (_log.isDebugEnabled()) {
+			_log.debug("Syndicating " + articles.size() + " articles");
+		}
+
+		for (JournalArticle article : articles) {
+			SyndEntry syndEntry = new SyndEntryImpl();
+
+			String author = PortalUtil.getUserName(article);
+
+			syndEntry.setAuthor(author);
+
+			SyndContent syndContent = new SyndContentImpl();
+
+			syndContent.setType(com.liferay.util.RSSUtil.ENTRY_TYPE_DEFAULT);
+
+			String value = article.getDescription(languageId);
+
+			try {
+				value = processContent(
+					feed, article, languageId, themeDisplay, syndEntry,
+					syndContent);
+			}
+			catch (Exception e) {
+				if (_log.isWarnEnabled()) {
+					_log.warn(e, e);
+				}
+			}
+
+			syndContent.setValue(value);
+
+			syndEntry.setDescription(syndContent);
+
+			String link = getEntryURL(
+				resourceRequest, feed, article, layout, themeDisplay);
+
+			syndEntry.setLink(link);
+
+			syndEntry.setPublishedDate(article.getDisplayDate());
+			syndEntry.setTitle(article.getTitle(languageId));
+			syndEntry.setUpdatedDate(article.getModifiedDate());
+			syndEntry.setUri(link);
+
+			syndEntries.add(syndEntry);
+		}
+
+		syndFeed.setFeedType(
+			feed.getFeedFormat() + "_" + feed.getFeedVersion());
+
+		List<SyndLink> syndLinks = new ArrayList<>();
+
+		syndFeed.setLinks(syndLinks);
+
+		SyndLink selfSyndLink = new SyndLinkImpl();
+
+		syndLinks.add(selfSyndLink);
+
+		ResourceURL feedURL = resourceResponse.createResourceURL();
+
+		feedURL.setCacheability(ResourceURL.FULL);
+		feedURL.setParameter("struts_action", "/journal/rss");
+		feedURL.setParameter("groupId", String.valueOf(feed.getGroupId()));
+		feedURL.setParameter("feedId", String.valueOf(feed.getFeedId()));
+
+		String link = feedURL.toString();
+
+		selfSyndLink.setHref(link);
+
+		selfSyndLink.setRel("self");
+
+		syndFeed.setTitle(feed.getName());
+		syndFeed.setPublishedDate(new Date());
+		syndFeed.setUri(feedURL.toString());
+
+		try {
+			return com.liferay.util.RSSUtil.export(syndFeed);
+		}
+		catch (FeedException fe) {
+			throw new SystemException(fe);
+		}
+	}
+
+	protected String getEntryURL(
+			ResourceRequest resourceRequest, JournalFeed feed,
+			JournalArticle article, Layout layout, ThemeDisplay themeDisplay)
+		throws Exception {
+
+		List<Long> hitLayoutIds =
+			JournalContentSearchLocalServiceUtil.getLayoutIds(
+				layout.getGroupId(), layout.isPrivateLayout(),
+				article.getArticleId());
+
+		if (!hitLayoutIds.isEmpty()) {
+			Long hitLayoutId = hitLayoutIds.get(0);
+
+			Layout hitLayout = LayoutLocalServiceUtil.getLayout(
+				layout.getGroupId(), layout.isPrivateLayout(),
+				hitLayoutId.longValue());
+
+			return PortalUtil.getLayoutFriendlyURL(hitLayout, themeDisplay);
+		}
+
+		String portletId = feed.getTargetPortletId();
+
+		if (Validator.isNull(portletId)) {
+			return StringPool.BLANK;
+		}
+
+		long plid = PortalUtil.getPlidFromFriendlyURL(
+			feed.getCompanyId(), feed.getTargetLayoutFriendlyUrl());
+
+		PortletURL entryURL = new PortletURLImpl(
+			resourceRequest, portletId, plid, PortletRequest.RENDER_PHASE);
+
+		entryURL.setParameter("groupId", String.valueOf(article.getGroupId()));
+		entryURL.setParameter("articleId", article.getArticleId());
+
+		return entryURL.toString();
+	}
+
+	@Override
+	protected byte[] getRSS(
+			ResourceRequest resourceRequest, ResourceResponse resourceResponse)
+		throws Exception {
+
+		ThemeDisplay themeDisplay = (ThemeDisplay)resourceRequest.getAttribute(
+			WebKeys.THEME_DISPLAY);
+
+		JournalFeed feed = null;
+
+		long id = ParamUtil.getLong(resourceRequest, "id");
+
+		long groupId = ParamUtil.getLong(resourceRequest, "groupId");
+		String feedId = ParamUtil.getString(resourceRequest, "feedId");
+
+		if (id > 0) {
+			feed = JournalFeedLocalServiceUtil.getFeed(id);
+		}
+		else {
+			feed = JournalFeedLocalServiceUtil.getFeed(groupId, feedId);
+		}
+
+		String languageId = LanguageUtil.getLanguageId(resourceRequest);
+
+		long plid = PortalUtil.getPlidFromFriendlyURL(
+			themeDisplay.getCompanyId(), feed.getTargetLayoutFriendlyUrl());
+
+		Layout layout = themeDisplay.getLayout();
+
+		if (plid > 0) {
+			try {
+				layout = LayoutLocalServiceUtil.getLayout(plid);
+			}
+			catch (NoSuchLayoutException nsle) {
+			}
+		}
+
+		String rss = exportToRSS(
+			resourceRequest, resourceResponse, feed, languageId, layout,
+			themeDisplay);
+
+		return rss.getBytes(StringPool.UTF8);
+	}
+
+	protected String processContent(
+			JournalFeed feed, JournalArticle article, String languageId,
+			ThemeDisplay themeDisplay, SyndEntry syndEntry,
+			SyndContent syndContent)
+		throws Exception {
+
+		String content = article.getDescription(languageId);
+
+		String contentField = feed.getContentField();
+
+		if (contentField.equals(JournalFeedConstants.RENDERED_WEB_CONTENT)) {
+			String ddmRendererTemplateKey = article.getDDMTemplateKey();
+
+			if (Validator.isNotNull(feed.getDDMRendererTemplateKey())) {
+				ddmRendererTemplateKey = feed.getDDMRendererTemplateKey();
+			}
+
+			JournalArticleDisplay articleDisplay =
+				JournalContentUtil.getDisplay(
+					feed.getGroupId(), article.getArticleId(),
+					ddmRendererTemplateKey, null, languageId, 1,
+					new PortletRequestModel() {
+
+						@Override
+						public String toXML() {
+							return _XML_REQUUEST;
+						}
+
+					},
+					themeDisplay);
+
+			if (articleDisplay != null) {
+				content = articleDisplay.getContent();
+			}
+		}
+		else if (!contentField.equals(
+					JournalFeedConstants.WEB_CONTENT_DESCRIPTION)) {
+
+			Document document = SAXReaderUtil.read(
+				article.getContentByLocale(languageId));
+
+			contentField = HtmlUtil.escapeXPathAttribute(contentField);
+
+			XPath xPathSelector = SAXReaderUtil.createXPath(
+				"//dynamic-element[@name=" + contentField + "]");
+
+			List<Node> results = xPathSelector.selectNodes(document);
+
+			if (results.isEmpty()) {
+				return content;
+			}
+
+			Element element = (Element)results.get(0);
+
+			String elType = element.attributeValue("type");
+
+			if (elType.equals("document_library")) {
+				String url = element.elementText("dynamic-content");
+
+				url = processURL(feed, url, themeDisplay, syndEntry);
+			}
+			else if (elType.equals("image") || elType.equals("image_gallery")) {
+				String url = element.elementText("dynamic-content");
+
+				url = processURL(feed, url, themeDisplay, syndEntry);
+
+				content =
+					content + "<br /><br /><img alt='' src='" +
+						themeDisplay.getURLPortal() + url + "' />";
+			}
+			else if (elType.equals("text_box")) {
+				syndContent.setType("text");
+
+				content = element.elementText("dynamic-content");
+			}
+			else {
+				content = element.elementText("dynamic-content");
+			}
+		}
+
+		return content;
+	}
+
+	protected String processURL(
+		JournalFeed feed, String url, ThemeDisplay themeDisplay,
+		SyndEntry syndEntry) {
+
+		url = StringUtil.replace(
+			url,
+			new String[] {
+				"@group_id@", "@image_path@", "@main_path@"
+			},
+			new String[] {
+				String.valueOf(feed.getGroupId()), themeDisplay.getPathImage(),
+				themeDisplay.getPathMain()
+			}
+		);
+
+		List<SyndEnclosure> syndEnclosures = com.liferay.portlet.journal.util.JournalRSSUtil.getDLEnclosures(
+				themeDisplay.getURLPortal(), url);
+
+		syndEnclosures.addAll(
+			com.liferay.portlet.journal.util.JournalRSSUtil.getIGEnclosures(themeDisplay.getURLPortal(), url));
+
+		syndEntry.setEnclosures(syndEnclosures);
+
+		List<SyndLink> syndLinks = com.liferay.portlet.journal.util.JournalRSSUtil.getDLLinks(
+				themeDisplay.getURLPortal(), url);
+
+		syndLinks.addAll(
+			com.liferay.portlet.journal.util.JournalRSSUtil.getIGLinks(themeDisplay.getURLPortal(), url));
+
+		syndEntry.setLinks(syndLinks);
+
+		return url;
+	}
+
+	private static final String _XML_REQUUEST =
+		"<request><parameters><parameter><name>rss</name><value>true</value>" +
+			"</parameter></parameters></request>";
+
+	private static final Log _log = LogFactoryUtil.getLog(JournalRSSUtil.class);
+
 }
